[rename] : #6 figma 와 name sync
